# Projeto1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Project API",
        "description": "Project API Information",
        "contact": {
            "name": "TPSI-DWB"
        },
        "servers": [
            "http://localhost:3000"
        ]
    },
    "paths": {
        "/product/views": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Sort products by Views",
                "description": "Returns the products by most viewed to least viewed",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/product/{id}/comments": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update products Image paths by ID",
                "description": "Returns the updated paths.",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Insert desired id",
                    "required": true,
                    "type": "string"
                }, {
                    "in": "body",
                    "name": "body",
                    "description": "Sample person",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Comments"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/product/{id}/images": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update products Image paths by ID",
                "description": "Returns the updated paths.",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Insert desired id",
                    "required": true,
                    "type": "string"
                }, {
                    "in": "body",
                    "name": "body",
                    "description": "Sample person",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Images"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/product/id": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Gets a product by ID",
                "description": "Returns the product from ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "query",
                    "name": "id",
                    "description": "gets the product from query ID",
                    "required": true
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/product/tags": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Gets a list of products with the same tag",
                "description": "Returns a list of products with the same tag",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "query",
                    "name": "tags",
                    "description": "gets products with same tag",
                    "required": true
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/product/{id}/incrementViews": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Increments Views",
                "description": "Returns Incremented result of Views.",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Insert desired id",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/seller": {
            "get": {
                "tags": [
                    "Seller"
                ],
                "summary": "Gets a list of products by seller",
                "description": "Returns a list of products by seller",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "query",
                    "name": "seller_id",
                    "description": "gets products from seller",
                    "required": true
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/product/{title}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Deletes a Product",
                "description": "Returns the id of the created product.",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "title",
                    "description": "Delete product",
                    "required": true
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Gets a list of products",
                "description": "Returns a list of products",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Creates product",
                "description": "Returns the id of the created product.",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Sample product",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Product"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
        "Comments": {
            "type": "string",
            "properties": {
                "comments": {
                    "type": "string"
                }
            }
        },
        "Images": {
            "type": "string",
            "properties": {
                "images": {
                    "type": "string"
                }
            }
        },
        "Product": {
            "type": "object",
            "properties": {
                "seller_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                },
                "images": {
                    "type": "string"
                },
                "comments": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Product"
            }
        }
    }
}